Split RuntimeShape into separate file to prepare for TFLM fork
PiperOrigin-RevId: 389024847
